a68: top-level misc files
This commit updates a few administrative files in the top-level
directory.

The MAINTAINERS file is updated with new entries for the Algol 68
front-end and the libga68 run-time library.

SECURITY.txt is updated to add libga68 to the list of the other
run-time libraries.

contrib/gcc-changelog/git_commit.py is updated so it knows about the
'algol68' bugzilla component.

Signed-off-by: Jose E. Marchesi <jemarch@gnu.org>

ChangeLog

	* MAINTAINERS: Add Algol 68 subsystems.
	* SECURITY.txt: add libga68 to list of libraries.

contrib/ChangeLog

	* gcc-changelog/git_commit.py: add algol68 bugzilla
	component.