a68: error rather than ICE on invalid Unicode points in string breaks
The string break '(...) is used to specify a list of characters having
some given unicode points, in the form of uXXXX or UXXXXXXXX.
However, not every four digits nor eight digits hexadecimal numbers
are actual valid Unicode codepoints.

The compiler checks for the validity of the specific codepoints, but
it was emitting an ICE in case these would not be valid ones.  This
patch changes this so now we emit a proper error like:

  /home/jemarch/foo.a68:2:25: error: invalid Unicode codepoint in string literal
      2 |       fputs (f, "     '(Uf09f94a5)");